Change font size and move Navbar to components folder
