+++ src/components/Navbar/style.css
@@ -1,0 +1,7 @@
+nav a {
+  @apply mr-4 font-semibold text-base text-gray-700 no-underline pb-2 w-24 block text-center;
+}
+
+.active {
+  @apply text-green-800 font-bold border-b-2 border-green-800;
+}
