+++ src/index.css
@@ -1,7 +1,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;800;900&display=swap');
 
 body {
-  @apply bg-gray-100 text-gray-700 antialiased;
+  @apply text-sm bg-gray-100 text-gray-700 antialiased;
   font-family: 'Inter', -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto',
     'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
@@ -12,7 +12,7 @@ a {
 }
 
 h2 {
-  @apply font-bold text-xl;
+  @apply font-bold text-lg;
 }
 
 .btn {
